refactor(client): restructure exports
