+++ packages/client/src/runtime/index.ts
@@ -11,8 +11,8 @@ export { DMMFHelper as DMMFClass } from './dmmf'
 export { type BaseDMMF, DMMF } from './dmmf-types'
 export type { PrismaClientOptions } from './getPrismaClient'
 export { getPrismaClient } from './getPrismaClient'
+export { objectEnumValues } from './object-enums'
 export { makeDocument, PrismaClientValidationError, transformDocument, unpack } from './query'
-export { AnyNull, DbNull, enumValues, JsonNull } from './symbol-enums'
 export type { DecimalJsLike } from './utils/decimalJsLike'
 export { findSync } from './utils/find'
 export { warnEnvConflicts } from './warnEnvConflicts'
